Fixed compiler warnings.
git-svn-id: https://svn.aros.org/svn/aros/trunk@18567 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/prefs/serial/locale.c
@@ -55,7 +55,7 @@ VOID LocalizeLabels( STRPTR * labels)
     STRPTR local;
     
     while (*labels) {
-        local = MSG(*labels);
+        local = MSG((ULONG) *labels);
         *labels++ = local;
     }
 }
